refactor: separate timestamp related utilities in time_util module
